// src/lib/cody/types.ts
export interface Board {
  id: string
  name: string
  type: 'label' | 'milestone' | 'all'
}